+++ apps/web/app/(auth-pages)/signup/page.tsx
@@ -1,5 +1,3 @@
-import { redirect } from 'next/navigation';
-
 import {
   SignupTitle,
   NameInput,
@@ -8,27 +6,14 @@ import {
   SignupButton,
   LoginInfo,
 } from '../../../components/signup';
-import { setServerSession } from '../../../services/login/setServerSession';
-import { createUserAccount } from '../../../services/signup/createUserAccount';
+import { signupAction } from '../../../server-actions/signup/signupAction';
 
 const SignupPage = () => {
-  async function handleSignup(formData: FormData) {
-    'use server';
-
-    const nickname = formData.get('nickname') as string;
-    const email = formData.get('email') as string;
-    const password = formData.get('password') as string;
-
-    await createUserAccount({ email, password, nickname });
-    await setServerSession(email, password);
-    redirect('/');
-  }
-
   return (
     <main className="flex min-h-screen items-center justify-center">
       <div className="w-full max-w-[375px] px-4">
         <SignupTitle />
-        <form action={handleSignup}>
+        <form action={signupAction}>
           <div className="space-y-4 pt-4 pb-12">
             <NameInput />
             <EmailInput />
